tools/PDFautogenerator/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-autogenerator"
version = "0.1.0"
description = "Benign one-page PDF autogenerator for research datasets."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "Pillow>=11.0.0",
  "PyMuPDF>=1.26.0",
  "PyYAML>=6.0.0",
  "pypdf>=6.0.0",
  "reportlab>=4.0.0",
]

[project.scripts]
pdf-autogenerator = "pdf_autogenerator.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py313"]
